Move typelits lemmas/proofs to separate package
We eventually intend to publish it on Hackage, so I didn't include the
custom data types / type classes: they remain in `bittide`.
