+++ bittide-instances/bittide-instances.cabal
@@ -108,6 +108,7 @@ common common-options
     extra,
     filepath,
     ghc-typelits-extra,
+    ghc-typelits-extra-lemmas,
     ghc-typelits-knownnat,
     ghc-typelits-natnormalise,
     lift-type,
@@ -227,6 +228,7 @@ test-suite unittests
     bittide-instances,
     clash-prelude-hedgehog >=1.6 && <1.10,
     filepath,
+    ghc-typelits-extra-lemmas,
     hedgehog >=1.0 && <1.5,
     parsec,
     pretty-simple,
