+++ ghc-typelits-extra-lemmas/README.md
@@ -1,0 +1,8 @@
+<!--
+SPDX-FileCopyrightText: 2025 Google LLC
+
+SPDX-License-Identifier: Apache-2.0
+-->
+# ghc-typelits-extra-lemmas
+
+Extra proofs related to type nats.
