+++ ghc-typelits-extra-lemmas/ghc-typelits-extra-lemmas.cabal.license
@@ -1,0 +1,3 @@
+SPDX-FileCopyrightText: 2025 Google LLC
+
+SPDX-License-Identifier: CC0-1.0
